refactor: major refactor of fee component + minor  adjustments | mz
