+++ src/hooks/index.ts
@@ -1,3 +1,2 @@
 export * from './useAxios';
-export * from './useBinance';
 export * from './useFees';
